requirements.txt
streamlit>=1.28.0
arxiv>=2.0.0
openai>=1.0.0
pymupdf>=1.23.0
wordcloud>=1.9.0
pillow>=10.0.0
